+++ src/client/js/functions.js
@@ -126,6 +126,9 @@ Modal.set = function( arg ){
     if( Object.keys( arg ).includes( 'footer' )){
         modal.footer( arg.footer );
     }
+    if( Object.keys( arg ).includes( 'fullscreen' )){
+        modal.fullScreen( arg.fullscreen );
+    }
     if( Object.keys( arg ).includes( 'target' )){
         modal.target( arg.target );
     }