+++ ChangeLog.md
@@ -22,6 +22,7 @@
     - Be verbose when resizing the modal (todo #17)
     - Let the dialog be displayed in full screen mode, an option which should be reserved to very special situation (todo #14)
     - Remove verticalPosition parameter, as it should rather be managed by a class
+    - Bump Meteor version requirement
 
 ### 1.6.0